Enabling explicit counter support in quick pulse module

+++ Src/PerformanceCollector/Shared/Implementation/IPerformanceCollector.cs
@@ -15,7 +15,7 @@
         /// Performs collection for all registered counters.
         /// </summary>
         /// <param name="onReadingFailure">Invoked when an individual counter fails to be read.</param>
-        IEnumerable<Tuple<PerformanceCounterData, float>> Collect(Action<string, Exception> onReadingFailure = null);
+        IEnumerable<Tuple<PerformanceCounterData, double>> Collect(Action<string, Exception> onReadingFailure = null);
 
         /// <summary>
         /// Refreshes and rebinds all the set of counters that are intended to be collected.
