+++ Src/PerformanceCollector/Shared/Implementation/StandardPerformanceCollector/StandardPerformanceCollector.cs
@@ -25,13 +25,13 @@
         /// Performs collection for all registered counters.
         /// </summary>
         /// <param name="onReadingFailure">Invoked when an individual counter fails to be read.</param>
-        public IEnumerable<Tuple<PerformanceCounterData, float>> Collect(
+        public IEnumerable<Tuple<PerformanceCounterData, double>> Collect(
             Action<string, Exception> onReadingFailure = null)
         {
             return this.performanceCounters.Where(pc => !pc.Item1.IsInBadState).SelectMany(
                 pc =>
                     {
-                        float value;
+                        double value;
 
                         try
                         {
@@ -46,7 +46,7 @@
                                     e);
                             }
 
-                            return new Tuple<PerformanceCounterData, float>[] { };
+                            return new Tuple<PerformanceCounterData, double>[] { };
                         }
 
                         return new[] { Tuple.Create(pc.Item1, value) };
@@ -103,7 +103,7 @@
         /// <summary>
         /// Collects a value for a single counter.
         /// </summary>
-        private static float CollectCounter(PerformanceCounter pc)
+        private static double CollectCounter(PerformanceCounter pc)
         {
             try
             {
