+++ Src/PerformanceCollector/Shared/Implementation/WebAppPerformanceCollector/SumUpCountersGauge.cs
@@ -35,7 +35,7 @@
         /// Returns the current value of the sum of all different gauges attached to this one and resets their values.
         /// </summary>
         /// <returns>The value of the target metric.</returns>
-        public float GetValueAndReset()
+        public double GetValueAndReset()
         {
             return this.gaugesToSum.Sum((g) => { return g.GetValueAndReset(); });
         }
