+++ Src/PerformanceCollector/Unit.Tests/WebAppPerformanceCollector/RateCounterTests.cs
@@ -11,7 +11,7 @@
         {
             RateCounterGauge privateBytesRate = new RateCounterGauge(@"\Process(??APP_WIN32_PROC??)\Private Bytes", "privateBytes", AzureWebApEnvironmentVariables.App, null, new CacheHelperTests());
 
-            float value = privateBytesRate.GetValueAndReset();
+            double value = privateBytesRate.GetValueAndReset();
             Assert.IsTrue(value == 0);
 
             System.Threading.Thread.Sleep(System.TimeSpan.FromSeconds(7));
