+++ Src/PerformanceCollector/Unit.Tests/WebAppPerformanceCollector/SumUpCountersGaugeTests.cs
@@ -16,8 +16,8 @@
 
             RawCounterGauge privateBytes = new RawCounterGauge(@"\Process(??APP_WIN32_PROC??)\Private Bytes", "privateBytes", AzureWebApEnvironmentVariables.App, new CacheHelperTests());
 
-            float expectedValue = privateBytes.GetValueAndReset();
-            float actualValue = twoTimesPrivateBytes.GetValueAndReset();
+            double expectedValue = privateBytes.GetValueAndReset();
+            double actualValue = twoTimesPrivateBytes.GetValueAndReset();
 
             // twoTimesPrivateBytes is -greater than (privateBytes * 1.85) but lower than (privateBytes * 2.15).
             Assert.IsTrue((expectedValue * 1.85) < actualValue && (expectedValue * 2.15) > actualValue);
